fix: restore build + tests after recent merges
- Resolve merge artifacts (scroll.rs) and duplicate definitions/imports
- Restore media module export and update call sites for changed APIs
- Refactor render_worker JS post-DOM-event pump to avoid self/tab borrow conflicts
- Fix borrow checker issues in audio drift, dom2 attribute syncing, and JS timer helpers
- Update tests for new PreparedPaintOptions/media provider, AudioClock, site isolation, and tungstenite APIs

This brings  and  (including navigation regression tests) back to green.

+++ src/js/vmjs/window_websocket.rs
@@ -5527,7 +5527,7 @@ mod tests {
     // below that limit.
     let msg_len = (MAX_WEBSOCKET_PENDING_EVENT_BYTES / 2).saturating_add(1);
     assert!(
-      msg_len <= MAX_WEBSOCKET_MESSAGE_BYTES as usize,
+      msg_len <= MAX_WEBSOCKET_MESSAGE_BYTES,
       "test message len must be <= MAX_WEBSOCKET_MESSAGE_BYTES"
     );
